requirements.txt
beautifulsoup4==4.11.1
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.5
filelock==3.8.0
gdown==4.5.1
huggingface-hub==0.0.12
idna==3.4
joblib==1.2.0
nlpaug==1.1.11
nltk==3.7
numpy==1.23.4
packaging==21.3
pandas==1.1.5
Pillow==9.2.0
protobuf==3.19.0
pyparsing==3.0.9
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2022.4
PyYAML==6.0
regex==2022.9.13
requests==2.28.1
sacremoses==0.0.53
scikit-learn==1.1.2
scipy==1.9.2
sentence-transformers==2.0.0
sentencepiece==0.1.97
six==1.16.0
soupsieve==2.3.2.post1
tensorboardX==2.4.1
threadpoolctl==3.1.0
tokenizers==0.10.3
torchaudio==0.12.1+cu116
torchvision==0.13.1
tqdm==4.64.1
transformers==4.8.1
typing_extensions==4.4.0
urllib3==1.26.12
